smart_mirror/requirements.txt
opencv-python
numpy
requests
python-dotenv
pyttsx3
mediapipe==0.10.14
Pillow
protobuf<5
pandas
matplotlib
PyQt5
solapi-python
supabase
gTTS
pygame